requirements.txt
pip
torch
transformers
smolagents[litellm,transformers,openai]
requests
markdownify
duckduckgo-search
wikipedia
pandas
pillow
accelerate
langchain-chroma
langchain-core
langchain-text-splitters
langchain-huggingface
chromadb
sentence-transformers
